:recycle: Refactoring classifier.js
'클린코드' 참고해서 classifier.js 리팩토링 진행, modules 하위 디렉토리 생성 및 파일 이동
