+++ modules/auth/addOauthToken.js
@@ -1,4 +1,4 @@
-const { users } = require('../models');
+const { users } = require('../../models');
 const jwt = require('jsonwebtoken');
 const dotenv = require('dotenv').config();
