+++ modules/auth/addToken.js
