+++ modules/auth/checkToken.js
