+++ modules/index.js
@@ -1,10 +1,12 @@
 module.exports = {
-  addToken: require('./addToken'),
-  checkToken: require('./checkToken'),
-  enToken: require('./enToken'),
-  classifier: require('./classifier'),
-  updateToken: require('./updateToken'),
-  addOauthToken: require('./addOauthToken'),
-  category: require('./category'),
-  cheerio: require('./cheerio'),
+  addToken: require('./auth/addToken'),
+  checkToken: require('./auth/checkToken'),
+  enToken: require('./auth/enToken'),
+  updateToken: require('./auth/updateToken'),
+  addOauthToken: require('./auth/addOauthToken'),
+  classifier: require('./classifier/classifier'),
+  category: require('./classifier/category'),
+  cheerio: require('./classifier/cheerio'),
+  handleException: require('./classifier/handleException'),
+  og: require('./og'),
 };
